Update README.md
Green's Lumber Sorter input and output. The user is to input the job order ID. They will then enter all of the pieces of wood that will be cut and shipped that that job order. After all pieces are input, the user will recieve a document file with all of the information they had provided in a printable and readable fashion.